# docs/04-api/openapi.yaml
openapi: 3.1.0
info:
  title: CNPM Digital Platform API
  version: 1.0.0
  description: Contrat API de référence pour les portails Web, mobile, back-office et intégrations CNPM.
servers:
- url: https://api.example.cnpm.ml/v1
  description: URL de production à confirmer
- url: https://api.sandbox.example.cnpm.ml/v1
  description: Sandbox à confirmer
security:
- oauth2:
  - openid
...
paths:
...
  /enrollment-applications:
    get:
      tags:
      - ENROLLMENT
      summary: Lister les dossiers d’enrôlement.
      operationId: listEnrollmentApplications
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Collection paginée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResource'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
    post:
      tags:
      - ENROLLMENT
      summary: Créer un dossier.
      operationId: createEnrollmentApplication
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Rejeu idempotent — dossier déjà existant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
        '201':
          description: Dossier créé au statut DRAFT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentApplicationInput'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
  /enrollment-applications/{id}:
    get:
      tags:
      - ENROLLMENT
      summary: Consulter un dossier.
      operationId: getEnrollmentApplication
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Ressource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
    patch:
      tags:
      - ENROLLMENT
      summary: Modifier un brouillon.
      operationId: updateEnrollmentApplication
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Opération effectuée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceInput'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
  /enrollment-applications/{id}/submit:
    post:
      tags:
      - ENROLLMENT
      summary: Soumettre le dossier.
      operationId: submitEnrollmentApplication
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Opération effectuée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
  /enrollment-applications/{id}/start-review:
    post:
      tags:
      - ENROLLMENT
      summary: Prendre le dossier en charge pour contrôle.
      description: >-
        Fait passer le dossier de SUBMITTED à UNDER_REVIEW et l'assigne au contrôleur. La
        machine à états normative (docs/07-processes/state-machines.md) impose le passage par
        UNDER_REVIEW avant toute décision : cette opération en est le seul point d'entrée,
        ce qui rend le contrôle (ENROLLMENT.REVIEW) prérequis de la décision
        (ENROLLMENT.APPROVE). Elle comble une opération absente du catalogue initial.
      operationId: startEnrollmentReview
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '200':
          description: Dossier pris en charge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
      x-requirements:
      - ENR-005
  /enrollment-applications/{id}/request-complement:
    post:
      tags:
      - ENROLLMENT
      summary: Demander un complément.
      operationId: requestEnrollmentComplement
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Opération effectuée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentComplementRequest'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
  /enrollment-applications/{id}/approve:
    post:
      tags:
      - ENROLLMENT
      summary: Approuver et activer.
      operationId: approveEnrollmentApplication
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Opération effectuée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentApprovalInput'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
  /enrollment-applications/{id}/reject:
    post:
      tags:
      - ENROLLMENT
      summary: Rejeter avec motif.
      operationId: rejectEnrollmentApplication
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '200':
          description: Opération effectuée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentApplicationView'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRejectionInput'
      x-requirements:
      - ENR-001
      - ENR-002
      - ENR-003
      - ENR-004
      - ENR-005
      - ENR-006
      - ENR-007
      - ENR-008
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://iam.example.cnpm.ml/realms/cnpm/protocol/openid-connect/auth
          tokenUrl: https://iam.example.cnpm.ml/realms/cnpm/protocol/openid-connect/token
          scopes:
            openid: Identité
            profile: Profil
            offline_access: Renouvellement contrôlé
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 100
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    Uuid:
      type: string
      format: uuid
...
    AuditFields:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        createdBy:
          $ref: '#/components/schemas/Uuid'
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    Resource:
      type: object
      required:
      - id
      - status
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        businessReference:
          type: string
        status:
          type: string
        attributes:
          type: object
          additionalProperties: true
        audit:
          $ref: '#/components/schemas/AuditFields'
    ResourceInput:
      type: object
      properties:
        businessReference:
          type: string
        attributes:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    PageResource:
      type: object
      required:
      - items
      - page
      - size
      - totalElements
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    EnrollmentApplicationInput:
      type: object
      description: >-
        Corps de création d'un dossier d'adhésion. Aucun contrôle de format n'est appliqué
        aux identifiants métier (ENR-003 différé : aucun masque RCCM/NIF n'est fourni par
        les sources) ; le statut initial DRAFT est imposé par le serveur.
      required:
      - caseNumber
      - organizationId
      - channel
      properties:
        caseNumber:
          type: string
          minLength: 1
          maxLength: 60
          description: Identité métier du dossier, unique — clé naturelle d'idempotence.
        organizationId:
          $ref: '#/components/schemas/Uuid'
        channel:
          type: string
          minLength: 1
          maxLength: 30
      additionalProperties: false
    EnrollmentApplicationView:
      type: object
      description: Dossier d'adhésion — état courant dans la machine à états d'enrôlement.
      required:
      - id
      - caseNumber
      - organizationId
      - channel
      - status
      - version
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        caseNumber:
          type: string
          maxLength: 60
        organizationId:
          $ref: '#/components/schemas/Uuid'
        channel:
          type: string
          maxLength: 30
        status:
          type: string
          enum: [DRAFT, SUBMITTED, UNDER_REVIEW, COMPLEMENT_REQUIRED, APPROVED, REJECTED]
        submittedAt:
          type: string
          format: date-time
          nullable: true
        assignedTo:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
          format: int64
      additionalProperties: false
    EnrollmentComplementRequest:
      type: object
      description: Demande de complément (ENR-005) — le motif est obligatoire.
      required:
      - comment
      properties:
        comment:
          type: string
          minLength: 1
          maxLength: 2000
      additionalProperties: false
    EnrollmentApprovalInput:
      type: object
      description: Corps d'une approbation d'enrôlement — le commentaire est facultatif.
      properties:
        comment:
          type: string
          maxLength: 2000
      additionalProperties: false
    EnrollmentRejectionInput:
      type: object
      description: >-
        Corps d'un rejet d'enrôlement. Le motif (comment) est OBLIGATOIRE : une décision
        défavorable non motivée n'est pas opposable au demandeur. reasonCode reste libre, le
        référentiel des motifs n'étant pas fourni par les sources (ENR-DEC-001).
      required:
      - comment
      properties:
        reasonCode:
          type: string
          maxLength: 60
        comment:
          type: string
          minLength: 1
          maxLength: 2000
      additionalProperties: false
...
    Problem:
      type: object
      required:
      - timestamp
      - status
      - code
      - message
      - correlationId
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
        correlationId:
          type: string
          format: uuid
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentification requise
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Permission ou périmètre insuffisant
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Ressource introuvable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflit d’état ou d’idempotence
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: Règle métier non satisfaite
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'